+++ .idea/compiler.xml
@@ -36,6 +36,7 @@
         </processorPath>
         <module name="core" />
         <module name="cmd-line" />
+        <module name="gateway-core" />
         <module name="aws" />
         <module name="gcp" />
       </profile>
@@ -44,6 +45,7 @@
       <module name="aws" target="11" />
       <module name="cmd-line" target="11" />
       <module name="core" target="11" />
+      <module name="gateway-core" target="11" />
       <module name="gcp" target="11" />
       <module name="impl" target="11" />
       <module name="root" target="11" />
