+++ java/core/pom.xml
@@ -51,6 +51,13 @@
             <version>${dependency.lombok.version}</version>
             <scope>provided</scope>
         </dependency>
+
+        <dependency>
+            <groupId>com.avaulta.gateway</groupId>
+            <artifactId>gateway-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+
         <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-lang3</artifactId>
@@ -137,7 +144,6 @@
             <artifactId>httpcore</artifactId>
             <version>${dependency.apache-httpcore.version}</version>
         </dependency>
-
 
 
         <dependency>
