+++ java/gateway/LICENSE.md
@@ -1,0 +1,14 @@
+# License
+
+Avaulta Gateway code is copyright 2022 Avaulta LLC, all rights reserved.
+
+Usage, modification, and distribution require a license from Avaulta LLC or reseller.
+
+
+## Worklytics
+
+Worklytics, Co. is a licensed reseller of Avaulta Gateway, authorized to publish or distribute the software in source or packaged form to its customers, including as a "white labeled" software.
+
+Any distribution or publication in source code form must retain this copyright and notice.
+
+
