+++ java/gateway/README.md
@@ -1,0 +1,5 @@
+# README
+
+Clone of source of Avaulta Gateway, to ease dependency/compilation. Note that everything here and under is governed by the license of Avaulta Gateway.
+
+
