+++ java/gateway/src/main/java/com/avaulta/gateway/pseudonyms/PseudonymizationStrategy.java
@@ -1,4 +1,4 @@
-package co.worklytics.psoxy;
+package com.avaulta.gateway.pseudonyms;
 
 import java.util.function.Function;
