+++ java/gateway/src/main/java/com/avaulta/gateway/pseudonyms/impl/PseudonymizationStrategyDelimImpl.java
@@ -1,32 +1,26 @@
-package co.worklytics.psoxy.impl;
+package com.avaulta.gateway.pseudonyms.impl;
 
-import co.worklytics.psoxy.PseudonymizationStrategy;
-import co.worklytics.psoxy.gateway.ConfigService;
-import co.worklytics.psoxy.gateway.ProxyConfigProperty;
-import com.google.common.base.Preconditions;
-import lombok.NoArgsConstructor;
+import com.avaulta.gateway.pseudonyms.PseudonymizationStrategy;
+
+import lombok.Getter;
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.commons.lang3.Validate;
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
-import javax.inject.Inject;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Base64;
 import java.util.function.Function;
 
-/**
- * impl of EncryptionStrategy, using delimiters to split hash + cryptotext
- */
-@NoArgsConstructor(onConstructor_ = @Inject)
-public class PseudonymizationStrategyDelimImpl implements PseudonymizationStrategy {
 
-    @Inject
-    ConfigService config;
-    SecretKeySpec encryptionKey;
+
+@RequiredArgsConstructor
+public class PseudonymizationStrategyDelimImpl implements PseudonymizationStrategy {
 
     static final String IV_DELIMITER = ":";
     static final int CIPHER_BLOCK_SIZE_BYTES = 16; //128-bits; AES uses 128-bit blocks, regardless of key-length
@@ -35,7 +29,10 @@ public class PseudonymizationStrategyDelimImpl implements PseudonymizationStrate
     Base64.Encoder encoder = Base64.getUrlEncoder().withoutPadding();
     Base64.Decoder decoder = Base64.getUrlDecoder();
 
-
+    @Getter
+    final String salt;
+    @Getter
+    final SecretKeySpec encryptionKey;
 
     @SneakyThrows
     Cipher getCipherInstance() {
@@ -69,7 +66,7 @@ public class PseudonymizationStrategyDelimImpl implements PseudonymizationStrate
     @Override
     public String getIdentifier(@NonNull String reversiblePseudonym) {
         String[] parts = reversiblePseudonym.split(IV_DELIMITER);
-        Preconditions.checkArgument(parts.length == 2, "Invalid ciphertext: %s", reversiblePseudonym);
+        Validate.validIndex(parts, 1, "Invalid ciphertext: %s", reversiblePseudonym);
         byte[] iv = extractIv(decoder.decode(parts[0]));
         byte[] cryptoText = decoder.decode(parts[1]);
 
@@ -78,19 +75,5 @@ public class PseudonymizationStrategyDelimImpl implements PseudonymizationStrate
 
         byte[] plain = cipher.doFinal(cryptoText);
         return new String(plain, StandardCharsets.UTF_8);
-    }
-
-    String getSalt() {
-        return config.getConfigPropertyOrError(ProxyConfigProperty.PSOXY_SALT);
-    }
-
-
-    SecretKeySpec getEncryptionKey() {
-        if (encryptionKey == null) {
-            //q: validate key length? we expect 256-bit (32 bytes)
-            String keyFromConfig = config.getConfigPropertyOrError(ProxyConfigProperty.PSOXY_ENCRYPTION_KEY);
-            encryptionKey = new SecretKeySpec(Base64.getDecoder().decode(keyFromConfig), "AES");
-        }
-        return encryptionKey;
     }
 }
