+++ java/gateway/src/main/java/com/avaulta/gateway/pseudonyms/impl/PseudonymizationStrategyImpl.java
@@ -1,28 +1,28 @@
-package co.worklytics.psoxy.impl;
+package com.avaulta.gateway.pseudonyms.impl;
 
-import co.worklytics.psoxy.PseudonymizationStrategy;
-import co.worklytics.psoxy.gateway.ConfigService;
-import co.worklytics.psoxy.gateway.ProxyConfigProperty;
-import lombok.NoArgsConstructor;
+import com.avaulta.gateway.pseudonyms.PseudonymizationStrategy;
+import lombok.Getter;
 import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.apache.commons.codec.digest.DigestUtils;
+
 
 import javax.crypto.Cipher;
 import javax.crypto.spec.IvParameterSpec;
 import javax.crypto.spec.SecretKeySpec;
-import javax.inject.Inject;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.Base64;
 import java.util.function.Function;
 
-@NoArgsConstructor(onConstructor_ = @Inject)
+@RequiredArgsConstructor
 public class PseudonymizationStrategyImpl implements PseudonymizationStrategy {
 
-    @Inject
-    ConfigService config;
-    SecretKeySpec encryptionKey;
+    @Getter
+    final String salt;
+    @Getter
+    final SecretKeySpec encryptionKey;
 
     static final int CIPHER_BLOCK_SIZE_BYTES = 16; //128-bits; AES uses 128-bit blocks, regardless of key-length
     static final int PSEUDONYM_SIZE_BYTES = 32; //SHA-256
@@ -99,18 +99,4 @@ public class PseudonymizationStrategyImpl implements PseudonymizationStrategy {
         return new String(plain, StandardCharsets.UTF_8);
     }
 
-    String getSalt() {
-        return config.getConfigPropertyOrError(ProxyConfigProperty.PSOXY_SALT);
-    }
-
-
-    SecretKeySpec getEncryptionKey() {
-        if (encryptionKey == null) {
-            //q: validate key length? we expect 256-bit (32 bytes)
-            //for flexibility, use password-based encryption key?
-            String keyFromConfig = config.getConfigPropertyOrError(ProxyConfigProperty.PSOXY_ENCRYPTION_KEY);
-            encryptionKey = new SecretKeySpec(Base64.getDecoder().decode(keyFromConfig), "AES");
-        }
-        return encryptionKey;
-    }
 }
