+++ java/gateway/src/test/java/com/avaulta/gateway/pseudonyms/impl/TestUtils.java
@@ -1,0 +1,21 @@
+package com.avaulta.gateway.pseudonyms.impl;
+
+import lombok.SneakyThrows;
+
+import javax.crypto.SecretKey;
+import javax.crypto.SecretKeyFactory;
+import javax.crypto.spec.PBEKeySpec;
+import javax.crypto.spec.SecretKeySpec;
+import java.security.spec.KeySpec;
+
+public class TestUtils {
+
+
+    @SneakyThrows
+    public static SecretKeySpec testKey() {
+        SecretKeyFactory factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA256");
+        KeySpec spec = new PBEKeySpec("secret".toCharArray(), "salt".getBytes(), 65536, 256);
+        SecretKey tmp = factory.generateSecret(spec);
+        return new SecretKeySpec(tmp.getEncoded(), "AES");
+    }
+}
